Condition "location_check" removal
* The "location_check" condition now gets removed, when used in combination with the "block_state_property" condition
* "getConditions" and "getFunctions" on the LootTableFile class are now extensions

(addresses #36)
